# docs/swagger.yaml
definitions:
...
  params.CreateScene:
    properties:
      scene:
        $ref: '#/definitions/table.Scene'
      sceneNode:
        items:
          $ref: '#/definitions/table.SceneNode'
        type: array
    type: object
...
  params.GetSceneList:
    properties:
      limit:
        type: integer
      name:
        description: 名称
        type: string
      offset:
        type: integer
    type: object
...
  table.Scene:
    properties:
      describe:
        description: 描述
        type: string
      global.UUID:
        type: string
      name:
        description: 话术名称
        type: string
      sceneNode:
        items:
          $ref: '#/definitions/table.SceneNode'
        type: array
    type: object
  table.SceneNode:
    properties:
      global.UUID:
        type: string
      name:
        description: 节点名称
        type: string
      sceneID:
        description: 话术ID
        type: string
      text:
        description: 文字
        type: string
    type: object
...
info:
  contact: {}
  description: pluto 永远滴神!!!
  title: Swagger pluto API
  version: 1.0.0
paths:
...
  /scene/createScene:
    post:
      parameters:
      - description: 新增话术
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/params.CreateScene'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 新增话术
      tags:
      - Scene
  /scene/getSceneInfoByID:
    get:
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 查询话术节点详情详情
      tags:
      - Scene
  /scene/getSceneList:
    get:
      parameters:
      - description: 查询话术列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/params.GetSceneList'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 查询话术列表
      tags:
      - Scene
  /scene/seleteScene:
    delete:
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 删除话术
      tags:
      - Scene
  /scene/setScene:
    put:
      parameters:
      - description: 修改话术
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/table.Scene'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 修改话术
      tags:
      - Scene
  /scene/setSceneNode:
    put:
      parameters:
      - description: 新增话术
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/table.SceneNode'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 修改话术节点
      tags:
      - Scene
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"